+++ changelog.md
@@ -1,6 +1,15 @@
 # Changelog
 
-## [1.10.1] - 2026-1-14 Admin & Notifications
+## [1.10.2] - 2026-1-14 Battlefield Unification
+
+### New Features
+- Tab view for the Battlefield
+- Submitted games page for logged in users
+
+### Improvements
+- Universal game filter form used across the site
+
+## [1.10.1] - 2026-1-13 Admin & Notifications
 
 ### New Features
 - Admin pending posts widget and management system